# homework4/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个博客API系统,包含用户注册登录、文章管理、评论管理等功能",
        "title": "Blog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:9527",
    "basePath": "/api/v1",
    "paths": {
        "/comment/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "为文章创建评论,需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "创建评论",
                "parameters": [
                    {
                        "description": "评论信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.CommentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/comment/list": {
            "get": {
                "description": "分页获取文章评论列表,不需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "评论管理"
                ],
                "summary": "获取评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "postId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/post/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建新文章,需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "创建文章",
                "parameters": [
                    {
                        "description": "文章信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.PostResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/post/delete": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除文章,需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "description": "文章ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.DeletePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/post/list": {
            "get": {
                "description": "分页获取文章列表,不需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "获取文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/post/update": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新文章信息,需要登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章管理"
                ],
                "summary": "更新文章",
                "parameters": [
                    {
                        "description": "文章信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.PostResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登录获取JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误或登录失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "创建新用户账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "service.CommentResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "评论内容",
                    "type": "string",
                    "example": "很棒的文章!"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "id": {
                    "description": "评论ID",
                    "type": "integer",
                    "example": 1
                },
                "postId": {
                    "description": "文章ID",
                    "type": "integer",
                    "example": 1
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "service.CreateCommentRequest": {
            "type": "object",
            "required": [
                "content",
                "postId"
            ],
            "properties": {
                "content": {
                    "description": "评论内容",
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "example": "很棒的文章!"
                },
                "postId": {
                    "description": "文章ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "service.CreatePostRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "example": "这是文章内容..."
                },
                "title": {
                    "description": "标题",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1,
                    "example": "我的第一篇文章"
                }
            }
        },
        "service.DeletePostRequest": {
            "type": "object",
            "required": [
                "postId"
            ],
            "properties": {
                "postId": {
                    "description": "文章ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "service.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码 必传",
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "description": "用户名 必传",
                    "type": "string",
                    "example": "testuser"
                }
            }
        },
        "service.LoginResponse": {
            "type": "object",
            "properties": {
                "nickname": {
                    "description": "昵称",
                    "type": "string",
                    "example": "测试用户"
                },
                "token": {
                    "description": "JWT 令牌",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "testuser"
                }
            }
        },
        "service.PostResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string",
                    "example": "这是文章内容..."
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "id": {
                    "description": "文章ID",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "标题",
                    "type": "string",
                    "example": "我的第一篇文章"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "userId": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "service.RegisterRequest": {
            "type": "object",
            "required": [
                "nickname",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "邮箱，可选，格式校验",
                    "type": "string",
                    "example": "test@example.com"
                },
                "nickname": {
                    "description": "昵称 必传，1-64个字符",
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1,
                    "example": "测试用户"
                },
                "password": {
                    "description": "密码 必传，至少6个字符",
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                },
                "username": {
                    "description": "用户名 必传，3-20个字符",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3,
                    "example": "testuser"
                }
            }
        },
        "service.UpdatePostRequest": {
            "type": "object",
            "required": [
                "postId"
            ],
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "example": "更新后的内容..."
                },
                "postId": {
                    "description": "文章ID",
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "description": "标题",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1,
                    "example": "更新后的标题"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "请输入JWT token,格式为Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}